Upgrade to 1.26.3

+++ mediawiki/includes/WatchedItem.php
@@ -1,34 +1,184 @@
 <?php
 /**
+ * Accessor and mutator for watchlist entries.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ * http://www.gnu.org/copyleft/gpl.html
+ *
  * @file
  * @ingroup Watchlist
  */
 
 /**
+ * Representation of a pair of user and title for watchlist entries.
+ *
  * @ingroup Watchlist
  */
 class WatchedItem {
-	var $mTitle, $mUser, $id, $ns, $ti;
+	/** @var Title */
+	public $mTitle;
+
+	/** @var User */
+	public $mUser;
+
+	/** @var int */
+	public $mCheckRights;
+
+	/** @var bool */
+	private $loaded = false;
+
+	/** @var bool */
+	private $watched;
+
+	/** @var string */
+	private $timestamp;
+
+	/**
+	 * Constant to specify that user rights 'editmywatchlist' and
+	 * 'viewmywatchlist' should not be checked.
+	 * @since 1.22
+	 */
+	const IGNORE_USER_RIGHTS = 0;
+
+	/**
+	 * Constant to specify that user rights 'editmywatchlist' and
+	 * 'viewmywatchlist' should be checked.
+	 * @since 1.22
+	 */
+	const CHECK_USER_RIGHTS = 1;
+
+	/**
+	 * Do DB master updates right now
+	 * @since 1.26
+	 */
+	const IMMEDIATE = 0;
+	/**
+	 * Do DB master updates via the job queue
+	 * @since 1.26
+	 */
+	const DEFERRED = 1;
 
 	/**
 	 * Create a WatchedItem object with the given user and title
-	 * @param $user User: the user to use for (un)watching
-	 * @param $title Title: the title we're going to (un)watch
-	 * @return WatchedItem object
+	 * @since 1.22 $checkRights parameter added
+	 * @param User $user The user to use for (un)watching
+	 * @param Title $title The title we're going to (un)watch
+	 * @param int $checkRights Whether to check the 'viewmywatchlist' and 'editmywatchlist' rights.
+	 *     Pass either WatchedItem::IGNORE_USER_RIGHTS or WatchedItem::CHECK_USER_RIGHTS.
+	 * @return WatchedItem
 	 */
-	public static function fromUserTitle( $user, $title ) {
+	public static function fromUserTitle( $user, $title,
+		$checkRights = WatchedItem::CHECK_USER_RIGHTS
+	) {
 		$wl = new WatchedItem;
 		$wl->mUser = $user;
 		$wl->mTitle = $title;
-		$wl->id = $user->getId();
-		# Patch (also) for email notification on page changes T.Gries/M.Arndt 11.09.2004
-		# TG patch: here we do not consider pages and their talk pages equivalent - why should we ?
-		# The change results in talk-pages not automatically included in watchlists, when their parent page is included
-		# $wl->ns = $title->getNamespace() & ~1;
-		$wl->ns = $title->getNamespace();
+		$wl->mCheckRights = $checkRights;
 
-		$wl->ti = $title->getDBkey();
 		return $wl;
+	}
+
+	/**
+	 * Title being watched
+	 * @return Title
+	 */
+	protected function getTitle() {
+		return $this->mTitle;
+	}
+
+	/**
+	 * Helper to retrieve the title namespace
+	 * @return int
+	 */
+	protected function getTitleNs() {
+		return $this->getTitle()->getNamespace();
+	}
+
+	/**
+	 * Helper to retrieve the title DBkey
+	 * @return string
+	 */
+	protected function getTitleDBkey() {
+		return $this->getTitle()->getDBkey();
+	}
+
+	/**
+	 * Helper to retrieve the user id
+	 * @return int
+	 */
+	protected function getUserId() {
+		return $this->mUser->getId();
+	}
+
+	/**
+	 * Return an array of conditions to select or update the appropriate database
+	 * row.
+	 *
+	 * @return array
+	 */
+	private function dbCond() {
+		return array(
+			'wl_user' => $this->getUserId(),
+			'wl_namespace' => $this->getTitleNs(),
+			'wl_title' => $this->getTitleDBkey(),
+		);
+	}
+
+	/**
+	 * Load the object from the database
+	 */
+	private function load() {
+		if ( $this->loaded ) {
+			return;
+		}
+		$this->loaded = true;
+
+		// Only loggedin user can have a watchlist
+		if ( $this->mUser->isAnon() ) {
+			$this->watched = false;
+			return;
+		}
+
+		// some pages cannot be watched
+		if ( !$this->getTitle()->isWatchable() ) {
+			$this->watched = false;
+			return;
+		}
+
+		# Pages and their talk pages are considered equivalent for watching;
+		# remember that talk namespaces are numbered as page namespace+1.
+
+		$dbr = wfGetDB( DB_SLAVE );
+		$row = $dbr->selectRow( 'watchlist', 'wl_notificationtimestamp',
+			$this->dbCond(), __METHOD__ );
+
+		if ( $row === false ) {
+			$this->watched = false;
+		} else {
+			$this->watched = true;
+			$this->timestamp = $row->wl_notificationtimestamp;
+		}
+	}
+
+	/**
+	 * Check permissions
+	 * @param string $what 'viewmywatchlist' or 'editmywatchlist'
+	 * @return bool
+	 */
+	private function isAllowed( $what ) {
+		return !$this->mCheckRights || $this->mUser->isAllowed( $what );
 	}
 
 	/**
@@ -36,47 +186,172 @@ class WatchedItem {
 	 * @return bool
 	 */
 	public function isWatched() {
-		# Pages and their talk pages are considered equivalent for watching;
-		# remember that talk namespaces are numbered as page namespace+1.
+		if ( !$this->isAllowed( 'viewmywatchlist' ) ) {
+			return false;
+		}
 
-		$dbr = wfGetDB( DB_SLAVE );
-		$res = $dbr->select( 'watchlist', 1, array( 'wl_user' => $this->id, 'wl_namespace' => $this->ns,
-			'wl_title' => $this->ti ), __METHOD__ );
-		$iswatched = ($dbr->numRows( $res ) > 0) ? 1 : 0;
-		return $iswatched;
+		$this->load();
+		return $this->watched;
 	}
 
 	/**
-	 * Given a title and user (assumes the object is setup), add the watch to the
-	 * database.
-	 * @return bool (always true)
+	 * Get the notification timestamp of this entry.
+	 *
+	 * @return bool|null|string False if the page is not watched, the value of
+	 *   the wl_notificationtimestamp field otherwise
+	 */
+	public function getNotificationTimestamp() {
+		if ( !$this->isAllowed( 'viewmywatchlist' ) ) {
+			return false;
+		}
+
+		$this->load();
+		if ( $this->watched ) {
+			return $this->timestamp;
+		} else {
+			return false;
+		}
+	}
+
+	/**
+	 * Reset the notification timestamp of this entry
+	 *
+	 * @param bool $force Whether to force the write query to be executed even if the
+	 *    page is not watched or the notification timestamp is already NULL.
+	 * @param int $oldid The revision id being viewed. If not given or 0, latest revision is assumed.
+	 * @mode int $mode WatchedItem::DEFERRED/IMMEDIATE
+	 */
+	public function resetNotificationTimestamp(
+		$force = '', $oldid = 0, $mode = self::IMMEDIATE
+	) {
+		// Only loggedin user can have a watchlist
+		if ( wfReadOnly() || $this->mUser->isAnon() || !$this->isAllowed( 'editmywatchlist' ) ) {
+			return;
+		}
+
+		if ( $force != 'force' ) {
+			$this->load();
+			if ( !$this->watched || $this->timestamp === null ) {
+				return;
+			}
+		}
+
+		$title = $this->getTitle();
+		if ( !$oldid ) {
+			// No oldid given, assuming latest revision; clear the timestamp.
+			$notificationTimestamp = null;
+		} elseif ( !$title->getNextRevisionID( $oldid ) ) {
+			// Oldid given and is the latest revision for this title; clear the timestamp.
+			$notificationTimestamp = null;
+		} else {
+			// See if the version marked as read is more recent than the one we're viewing.
+			// Call load() if it wasn't called before due to $force.
+			$this->load();
+
+			if ( $this->timestamp === null ) {
+				// This can only happen if $force is enabled.
+				$notificationTimestamp = null;
+			} else {
+				// Oldid given and isn't the latest; update the timestamp.
+				// This will result in no further notification emails being sent!
+				$notificationTimestamp = Revision::getTimestampFromId( $title, $oldid );
+				// We need to go one second to the future because of various strict comparisons
+				// throughout the codebase
+				$ts = new MWTimestamp( $notificationTimestamp );
+				$ts->timestamp->add( new DateInterval( 'PT1S' ) );
+				$notificationTimestamp = $ts->getTimestamp( TS_MW );
+
+				if ( $notificationTimestamp < $this->timestamp ) {
+					if ( $force != 'force' ) {
+						return;
+					} else {
+						// This is a little silly…
+						$notificationTimestamp = $this->timestamp;
+					}
+				}
+			}
+		}
+
+		// If the page is watched by the user (or may be watched), update the timestamp
+		if ( $mode === self::DEFERRED ) {
+			$job = new ActivityUpdateJob(
+				$title,
+				array(
+					'type'      => 'updateWatchlistNotification',
+					'userid'    => $this->getUserId(),
+					'notifTime' => $notificationTimestamp,
+					'curTime'   => time()
+				)
+			);
+			// Try to run this post-send
+			DeferredUpdates::addCallableUpdate( function() use ( $job ) {
+				$job->run();
+			} );
+		} else {
+			$dbw = wfGetDB( DB_MASTER );
+			$dbw->update( 'watchlist',
+				array( 'wl_notificationtimestamp' => $notificationTimestamp ),
+				$this->dbCond(),
+				__METHOD__
+			);
+		}
+
+		$this->timestamp = null;
+	}
+
+	/**
+	 * @param WatchedItem[] $items
+	 * @return bool
+	 */
+	public static function batchAddWatch( array $items ) {
+
+		if ( wfReadOnly() ) {
+			return false;
+		}
+
+		$rows = array();
+		foreach ( $items as $item ) {
+			// Only loggedin user can have a watchlist
+			if ( $item->mUser->isAnon() || !$item->isAllowed( 'editmywatchlist' ) ) {
+				continue;
+			}
+			$rows[] = array(
+				'wl_user' => $item->getUserId(),
+				'wl_namespace' => MWNamespace::getSubject( $item->getTitleNs() ),
+				'wl_title' => $item->getTitleDBkey(),
+				'wl_notificationtimestamp' => null,
+			);
+			// Every single watched page needs now to be listed in watchlist;
+			// namespace:page and namespace_talk:page need separate entries:
+			$rows[] = array(
+				'wl_user' => $item->getUserId(),
+				'wl_namespace' => MWNamespace::getTalk( $item->getTitleNs() ),
+				'wl_title' => $item->getTitleDBkey(),
+				'wl_notificationtimestamp' => null
+			);
+			$item->watched = true;
+		}
+
+		if ( !$rows ) {
+			return false;
+		}
+
+		$dbw = wfGetDB( DB_MASTER );
+		foreach ( array_chunk( $rows, 100 ) as $toInsert ) {
+			// Use INSERT IGNORE to avoid overwriting the notification timestamp
+			// if there's already an entry for this page
+			$dbw->insert( 'watchlist', $toInsert, __METHOD__, 'IGNORE' );
+		}
+
+		return true;
+	}
+
+	/**
+	 * Given a title and user (assumes the object is setup), add the watch to the database.
+	 * @return bool
 	 */
 	public function addWatch() {
-		wfProfileIn( __METHOD__ );
-
-		// Use INSERT IGNORE to avoid overwriting the notification timestamp
-		// if there's already an entry for this page
-		$dbw = wfGetDB( DB_MASTER );
-		$dbw->insert( 'watchlist',
-		  array(
-			'wl_user' => $this->id,
-			'wl_namespace' => MWNamespace::getSubject($this->ns),
-			'wl_title' => $this->ti,
-			'wl_notificationtimestamp' => null
-		  ), __METHOD__, 'IGNORE' );
-
-		// Every single watched page needs now to be listed in watchlist;
-		// namespace:page and namespace_talk:page need separate entries:
-		$dbw->insert( 'watchlist',
-		  array(
-			'wl_user' => $this->id,
-			'wl_namespace' => MWNamespace::getTalk($this->ns),
-			'wl_title' => $this->ti,
-			'wl_notificationtimestamp' => null
-		  ), __METHOD__, 'IGNORE' );
-
-		wfProfileOut( __METHOD__ );
-		return true;
+		return self::batchAddWatch( array( $this ) );
 	}
 
 	/**
@@ -84,30 +359,34 @@ class WatchedItem {
 	 * @return bool
 	 */
 	public function removeWatch() {
-		wfProfileIn( __METHOD__ );
+
+		// Only loggedin user can have a watchlist
+		if ( wfReadOnly() || $this->mUser->isAnon() || !$this->isAllowed( 'editmywatchlist' ) ) {
+			return false;
+		}
 
 		$success = false;
 		$dbw = wfGetDB( DB_MASTER );
 		$dbw->delete( 'watchlist',
 			array(
-				'wl_user' => $this->id,
-				'wl_namespace' => MWNamespace::getSubject($this->ns),
-				'wl_title' => $this->ti
+				'wl_user' => $this->getUserId(),
+				'wl_namespace' => MWNamespace::getSubject( $this->getTitleNs() ),
+				'wl_title' => $this->getTitleDBkey(),
 			), __METHOD__
 		);
 		if ( $dbw->affectedRows() ) {
 			$success = true;
 		}
 
-		# the following code compensates the new behaviour, introduced by the
+		# the following code compensates the new behavior, introduced by the
 		# enotif patch, that every single watched page needs now to be listed
 		# in watchlist namespace:page and namespace_talk:page had separate
 		# entries: clear them
 		$dbw->delete( 'watchlist',
 			array(
-				'wl_user' => $this->id,
-				'wl_namespace' => MWNamespace::getTalk($this->ns),
-				'wl_title' => $this->ti
+				'wl_user' => $this->getUserId(),
+				'wl_namespace' => MWNamespace::getTalk( $this->getTitleNs() ),
+				'wl_title' => $this->getTitleDBkey(),
 			), __METHOD__
 		);
 
@@ -115,7 +394,8 @@ class WatchedItem {
 			$success = true;
 		}
 
-		wfProfileOut( __METHOD__ );
+		$this->watched = false;
+
 		return $success;
 	}
 
@@ -123,8 +403,8 @@ class WatchedItem {
 	 * Check if the given title already is watched by the user, and if so
 	 * add watches on a new title. To be used for page renames and such.
 	 *
-	 * @param $ot Title: page title to duplicate entries from, if present
-	 * @param $nt Title: page title to add watches on
+	 * @param Title $ot Page title to duplicate entries from, if present
+	 * @param Title $nt Page title to add watches on
 	 */
 	public static function duplicateEntries( $ot, $nt ) {
 		WatchedItem::doDuplicateEntries( $ot->getSubjectPage(), $nt->getSubjectPage() );
@@ -134,19 +414,20 @@ class WatchedItem {
 	/**
 	 * Handle duplicate entries. Backend for duplicateEntries().
 	 *
-	 * @param $ot Title
-	 * @param $nt Title
+	 * @param Title $ot
+	 * @param Title $nt
 	 *
 	 * @return bool
 	 */
-	private static function doDuplicateEntries( $ot, $nt ) {	
+	private static function doDuplicateEntries( $ot, $nt ) {
 		$oldnamespace = $ot->getNamespace();
 		$newnamespace = $nt->getNamespace();
 		$oldtitle = $ot->getDBkey();
 		$newtitle = $nt->getDBkey();
 
 		$dbw = wfGetDB( DB_MASTER );
-		$res = $dbw->select( 'watchlist', 'wl_user',
+		$res = $dbw->select( 'watchlist',
+			array( 'wl_user', 'wl_notificationtimestamp' ),
 			array( 'wl_namespace' => $oldnamespace, 'wl_title' => $oldtitle ),
 			__METHOD__, 'FOR UPDATE'
 		);
@@ -156,11 +437,12 @@ class WatchedItem {
 			$values[] = array(
 				'wl_user' => $s->wl_user,
 				'wl_namespace' => $newnamespace,
-				'wl_title' => $newtitle
+				'wl_title' => $newtitle,
+				'wl_notificationtimestamp' => $s->wl_notificationtimestamp,
 			);
 		}
 
-		if( empty( $values ) ) {
+		if ( empty( $values ) ) {
 			// Nothing to do
 			return true;
 		}
@@ -168,7 +450,13 @@ class WatchedItem {
 		# Perform replace
 		# Note that multi-row replace is very efficient for MySQL but may be inefficient for
 		# some other DBMSes, mostly due to poor simulation by us
-		$dbw->replace( 'watchlist', array( array( 'wl_user', 'wl_namespace', 'wl_title' ) ), $values, __METHOD__ );
+		$dbw->replace(
+			'watchlist',
+			array( array( 'wl_user', 'wl_namespace', 'wl_title' ) ),
+			$values,
+			__METHOD__
+		);
+
 		return true;
 	}
 }
